requirements.txt
psycopg2-binary
flask 
flask_sqlalchemy
flask_migrate


